backend/requirements.txt
# Core framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Market data
yfinance>=0.2.28
alpaca-trade-api>=3.1.1

# Analysis
scipy>=1.11.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
pyyaml>=6.0.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For FastAPI testing
factory-boy>=3.3.0  # Test data factories
faker>=20.1.0  # Generate realistic test data

# Code Quality
mypy>=1.7.0
pylint>=3.0.0
black>=23.0.0
isort>=5.12.0
pre-commit>=3.5.0

# Monitoring & Debugging
pydantic-extra-types>=2.1.0  # Enhanced validation
structlog>=23.2.0  # Already there
sentry-sdk[fastapi]>=1.38.0  # Error tracking

# Logging
structlog>=23.2.0

# Scheduling
apscheduler>=3.10.4